feat(reader): Show current title and chapter in ui

+++ crates/illustrator/src/cache.rs
@@ -32,9 +32,9 @@ impl Default for PageContentCache {
 
 pub struct PageMetadata {
 	/// Page number, starting from 1
-	pub page: u64,
+	pub page: u32,
 	/// Total number of pages for chapter
-	pub pages: u64,
+	pub pages: u32,
 }
 
 #[derive(Debug, thiserror::Error)]
@@ -172,7 +172,7 @@ impl PageContentCache {
 		let (page, meta) = if loc.element == I26F6::ZERO {
 			let meta = PageMetadata {
 				page: 1,
-				pages: entry.pages.len() as u64,
+				pages: entry.pages.len() as u32,
 			};
 			(entry.pages.first()?, meta)
 		} else {
@@ -183,8 +183,8 @@ impl PageContentCache {
 				.find(|(_, p)| p.elements.contains(&loc.element))
 				.or_else(|| entry.pages.last().map(|p| (entry.pages.len() - 1, p)))?;
 			let meta = PageMetadata {
-				page: index as u64 + 1,
-				pages: entry.pages.len() as u64,
+				page: index as u32 + 1,
+				pages: entry.pages.len() as u32,
 			};
 			(page, meta)
 		};
